Ingestion_and_Parsing/requirements.txt
python-dotenv
requests
tqdm
supabase
pymupdf
pymupdf4llm
openai
tiktoken
pyyaml
neo4j
pinecone
sentry-sdk